HE MOVES!

+++ src/sdp/ggj14/game/Projectile.java
@@ -2,7 +2,7 @@ package sdp.ggj14.game;
 
 public class Projectile extends Unit {
 	
-	public Projectile(int x, int y, int hp) {
+	public Projectile(double x, double y, int hp) {
 		super(x, y, hp);
 		super.sprite = "/projectiles/player_default.png";
 	}
